# packages/db/OPENAPI.yaml
openapi: 3.0.3
info: { title: Beauty Booking API, version: 1.0.0 }
servers: [ { url: https://your-railway-app.up.railway.app } ]
paths:
...
  /api/masters:
    get:
      summary: List masters
      parameters: [ { in: query, name: service_id, schema: { type: string } } ]
      responses: { '200': { description: OK } }
...
  /api/appointments:
    post:
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, master_id, service_id, start_at]
              properties:
                client_id: { type: string }
                master_id: { type: string }
                service_id: { type: string }
                start_at: { type: string, format: date-time }
                comment: { type: string }
      responses: { '201': { description: Created } }
...